Rewrote most of the game package.
Adding correct handling for ring, no limit, and other game protocols should be
more straightforward now. Also, the player can now observe the game state when
it is not his turn to act.

For some reason the 7cHS agent is now doing worse against the acpc r/c dummy
now. It is likely related to when "roundActions" was telling the agent to
re-calculate the hand equity, but it is also possible the pot, call, and/or
raise sizes are not equal to what they were before.