Fixed the 7cHS strategy regression.
The number of raises was not being reset at the correct place incorrectly
limiting the number of raises the agent could make. Fixing this not only got
rid of the regression, but made the agent about twice as strong as the previous
version in which the agent was updating the hand equity only half as often as
it was supposed to be.